pathchirp, available bandwidth estimation,ns2
PathChirp is a well-known tool to estimate available bandwidth based on
an approach of self-induced congestion, which exponentially increases
the rate of probing packets in each packet train. But the inappropriate
gaps of probing rates often result in a large deviation of estimated
available bandwidth. In order to solve this problem, we propose
HybChirp, an active probing tool which uses linearly spaced probing
packets (lin- chirps) and exponentially spaced probing packets
(exp-chirps) in combination. HybChirp improves PathChirp by adopting the
hybrid of lin-chirps and exp-chirps, while inherits the advantages of
PathChirp such as the short convergence time and low overhead. The NS-2
simulation results have shown that HybChirp outperforms PathChirp and
Pathload in terms of accuracy, overhead, convergence time and
adaptivity.